+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### 1.4.3
 - Storage Monitors don't render any quantity text when no item is specified to monitor anymore (raoulvdberge)
-- The Solderer isn't sided anymore (raoulvdberge)
+- Fixed bug where disks in Disk Drive didn't respect access type or void excess stacks option (raoulvdberge)
+- The Solderer inventory isn't sided anymore (raoulvdberge)
 
 ### 1.4.2
 - Updated Forge to 2261 (raoulvdberge)